+++ src/metajumper/index.ts
@@ -1,7 +1,7 @@
 import { Config } from "../config";
 import { Utilities } from '../lib';
-import { InlineInput } from "./inline-input";
-import { ILineCharIndexes, IIndexes, DecorationModel, DecorationModelBuilder, InteliAdjustment, Direction, CharIndex } from "./decoration-model";
+import { Input } from "./input";
+import { ILineCharIndexes, IIndexes, DecorationModel, DecorationModelBuilder, SmartAdjustment, Direction, CharIndex } from "./decoration-model";
 import { Decorator } from "./decoration";
 import { ViewPort } from '../lib/viewport';
 import * as vscode from "vscode";
@@ -24,7 +24,7 @@ export class MetaJumper {
     private findFromCenterScreenRange;
     private halfViewPortRange: number;
     private currentFindIndex = -1;
-    private decorationModels: DecorationModel[];
+    private editorToDecorationModels = new Map<vscode.TextEditor, DecorationModel[]>()
     private isSelectionMode: boolean;
 
     constructor(context: vscode.ExtensionContext, config: Config) {
@@ -46,6 +46,19 @@ export class MetaJumper {
 
         disposables.push(vscode.commands.registerCommand('metaGo.gotoBefore', () => {
             this.jumpTo(JumpPosition.Before);
+        }));
+
+        disposables.push(vscode.commands.registerCommand('metaGo.gotoAfterActive', () => {
+            this.jumpTo(JumpPosition.After, false);
+        }));
+
+        disposables.push(vscode.commands.registerCommand('metaGo.gotoSmartActive', () => {
+            this.jumpTo(JumpPosition.Smart, false)
+
+        }));
+
+        disposables.push(vscode.commands.registerCommand('metaGo.gotoBeforeActive', () => {
+            this.jumpTo(JumpPosition.Before, false);
         }));
 
         disposables.push(vscode.commands.registerCommand('metaGo.selectSmart', () => {
@@ -72,13 +85,12 @@ export class MetaJumper {
         this.decorator.initialize(this.config);
     }
 
-    private async jumpTo(jumpPosition: JumpPosition) {
+    private async jumpTo(jumpPosition: JumpPosition, mutiEditor: boolean = true) {
         this.isSelectionMode = false;
 
         try {
-            var model = await this.getLocationWithTimeout()
+            var [editor, model] = await this.getLocationWithTimeout(mutiEditor)
             this.done();
-            let editor = vscode.window.activeTextEditor;
             switch (jumpPosition) {
                 case JumpPosition.Before:
                     Utilities.goto(editor, model.lineIndex, model.charIndex);
@@ -110,7 +122,7 @@ export class MetaJumper {
         let fromLine = position.line;
         let fromChar = position.character;
         try {
-            var model = await this.getLocationWithTimeout()
+            var [ed, model] = await this.getLocationWithTimeout(false)
             this.done();
             let toCharacter = model.charIndex;
 
@@ -147,8 +159,8 @@ export class MetaJumper {
     }
 
     private cancel() {
-        while (InlineInput.instances.length > 0) {
-            InlineInput.instances[0].cancelInput();
+        while (Input.instances.length > 0) {
+            Input.instances[0].cancelInput();
         }
         this.isJumping = false;
     }
@@ -167,13 +179,13 @@ export class MetaJumper {
     }
     private jumpTimeoutId = null;
 
-    private async getLocationWithTimeout() {
+    private async getLocationWithTimeout(mutiEditor: boolean) {
         if (!this.isJumping) {
             this.isJumping = true;
 
             this.jumpTimeoutId = setTimeout(() => { this.jumpTimeoutId = null; this.cancel(); }, this.config.jumper.timeout);
             try {
-                var [, model] = await this.getLocation();
+                var model = await this.getLocation(mutiEditor);
                 return model;
             }
             finally {
@@ -187,60 +199,73 @@ export class MetaJumper {
         }
     }
 
-    private async getLocation(): Promise<[vscode.TextEditor, DecorationModel]> {
-        let editor = vscode.window.activeTextEditor;
-        if (!editor) {
-            throw new Error('no active editor');
+    private async getLocation(mutiEditor: boolean): Promise<[vscode.TextEditor, DecorationModel]> {
+        let inputEditor = vscode.window.activeTextEditor ?? vscode.window.visibleTextEditors[0];
+        if (!inputEditor) {
+            throw new Error('no visible editor');
         }
 
         let msg = this.isSelectionMode ? "metaGo: Type to Select" : "metaGo: Type To Jump"
         let messageDisposable = vscode.window.setStatusBarMessage(msg, this.config.jumper.timeout);
 
         try {
-            this.decorator.addCommandIndicator(editor);
+            this.decorator.addCommandIndicator(inputEditor);
 
-            var locationChars = await this.getLocationChar(editor);
+            var locationChars = await this.getLocationChar(inputEditor);
             if (!locationChars) {
                 throw new Error('no location char input')
             }
             if (locationChars && locationChars.length > 1)
                 locationChars = locationChars.substring(0, 1);
 
+            var editor: vscode.TextEditor = null;
+            var models: DecorationModel[] = null;
             var model: DecorationModel = null;
-            if (locationChars === ' ' && this.currentFindIndex !== Number.NaN && this.decorationModels) {
-                let model = this.decorationModels.find((model) => model.indexInModels === (this.currentFindIndex + 1));
-                if (model) {
-                    this.currentFindIndex++;
-                } else {
-                    throw new Error('metaGo: no next find');
-                }
-            } else if (locationChars === '\n' && this.currentFindIndex !== Number.NaN && this.decorationModels) {
-                let model = this.decorationModels.find((model) => model.indexInModels === (this.currentFindIndex - 1));
-                if (model) {
-                    this.currentFindIndex--;
-                } else {
-                    throw new Error('metaGo: no previous find');
-                }
-            } else {
-                var selection = await this.getJumpRange(editor);
-                let lineCharIndexes = this.find(editor, selection.before, selection.after, locationChars);
-                if (lineCharIndexes.count <= 0) {
-                    throw new Error("metaGo: no matches");
-                }
-
-                this.decorationModels = this.decorationModelBuilder.buildDecorationModel(lineCharIndexes);
-
-                var models = this.decorationModels;
-                if (models.length === 0) {
-                    throw new Error("metaGo: encoding error")
-                }
-                while (models.length > 1) {
-                    models = await this.getExactLocation(editor, models);
-                }
-                model = models[0]; // only one, length == 1
-                this.currentFindIndex = model.indexInModels;
-
+            // if (locationChars === ' ' && this.currentFindIndex !== Number.NaN && this.editorToDecorationModels) {
+            //     let model = this.editorToDecorationModels.find(model => model.indexInModels === (this.currentFindIndex + 1));
+            //     if (model) {
+            //         this.currentFindIndex++;
+            //     } else {
+            //         throw new Error('metaGo: no next find');
+            //     }
+            // } else if (locationChars === '\n' && this.currentFindIndex !== Number.NaN && this.editorToDecorationModels) {
+            //     let model = this.editorToDecorationModels.find(model => model.indexInModels === (this.currentFindIndex - 1));
+            //     if (model) {
+            //         this.currentFindIndex--;
+            //     } else {
+            //         throw new Error('metaGo: no previous find');
+            //     }
+            // } else
+            //{
+            let editorToLineCharIndexesMap = new Map<vscode.TextEditor, ILineCharIndexes>()
+            let editors = mutiEditor ? [inputEditor, ...vscode.window.visibleTextEditors.filter(e => e !== inputEditor)] : [inputEditor]
+            for (let editor of editors) {
+                var jumpRange = await this.getJumpRange(editor);
+                let lineCharIndexes = this.find(editor, jumpRange.before, jumpRange.after, locationChars);
+                if (lineCharIndexes.count > 0)
+                    editorToLineCharIndexesMap.set(editor, lineCharIndexes);
             }
+
+            let locationCount = 0;
+            editorToLineCharIndexesMap.forEach(lineCharIndex => locationCount += lineCharIndex.count)
+            if (locationCount <= 0) {
+                throw new Error("metaGo: no location match for input char");
+            }
+
+            let editorToModelsMap = this.decorationModelBuilder.buildDecorationModel(editorToLineCharIndexesMap, locationCount);
+            this.editorToDecorationModels = editorToModelsMap;
+            // here, we have editorToModelsMap.size > 1 || models.length > 1
+            do {
+                editorToModelsMap = await this.getExactLocation(editorToModelsMap);
+
+                if (editorToModelsMap.size == 0) throw new Error("metaGo: no match in codes for input char");
+
+                [editor, models] = editorToModelsMap.entries().next().value; // first entry
+            } while (editorToModelsMap.size > 1 || models.length > 1);
+
+            model = models[0];
+            this.currentFindIndex = model.indexInModels;
+            //}
 
             let msg = this.isSelectionMode ? 'metaGo: Selected!' : 'metaGo: Jumped!';
             vscode.window.setStatusBarMessage(msg, 2000);
@@ -248,7 +273,7 @@ export class MetaJumper {
 
         } catch (reason) {
             if (!reason) reason = new Error("Canceled!");
-            this.decorator.removeCommandIndicator(editor);
+            this.decorator.removeCommandIndicator(inputEditor);
             vscode.window.setStatusBarMessage(`metaGo: ${reason}`, 2000);
             throw reason;
         } finally {
@@ -257,7 +282,7 @@ export class MetaJumper {
     }
 
     private getLocationChar = (editor: vscode.TextEditor) => {
-        let result = new InlineInput(this.config)
+        let result = new Input(this.config)
             .input(editor, (v) => {
                 this.decorator.removeCommandIndicator(editor);
                 return v;
@@ -298,7 +323,7 @@ export class MetaJumper {
     }
 
     private find = (editor: vscode.TextEditor, selectionBefore: Selection, selectionAfter: Selection, locationChars: string): ILineCharIndexes => {
-        let lineCharIndexes: ILineCharIndexes = {
+        let editorLineCharIndexes: ILineCharIndexes = {
             count: 0,
             focusLine: 0,
             indexes: {}
@@ -307,40 +332,40 @@ export class MetaJumper {
         for (let i = selectionBefore.startLine; i < selectionBefore.lastLine; i++) {
             let line = editor.document.lineAt(i);
             let indexes = this.indexesOf(line.text, locationChars);
-            lineCharIndexes.count += indexes.length;
-            lineCharIndexes.indexes[i] = indexes;
+            editorLineCharIndexes.count += indexes.length;
+            editorLineCharIndexes.indexes[i] = indexes;
         }
-        lineCharIndexes.focusLine = editor.selection.active.line;
+        editorLineCharIndexes.focusLine = editor.selection.active.line;
 
         for (let i = selectionAfter.startLine; i < selectionAfter.lastLine; i++) {
             let line = editor.document.lineAt(i);
             let indexes = this.indexesOf(line.text, locationChars);
-            lineCharIndexes.count += indexes.length;
-            lineCharIndexes.indexes[i] = indexes;
+            editorLineCharIndexes.count += indexes.length;
+            editorLineCharIndexes.indexes[i] = indexes;
         }
-        return lineCharIndexes;
+        return editorLineCharIndexes;
     }
 
-    private smartAdjBefore(str: string, char: string, index: number): InteliAdjustment {
+    private smartAdjBefore(str: string, char: string, index: number): SmartAdjustment {
         let regexp = new RegExp('\\w');
-        if (this.smartAdjBeforeRegex(str, char, index, regexp) === InteliAdjustment.Before) {
-            return InteliAdjustment.Before;
+        if (this.smartAdjBeforeRegex(str, char, index, regexp) === SmartAdjustment.Before) {
+            return SmartAdjustment.Before;
         }
         regexp = new RegExp(/[^\w\s]/);
         return this.smartAdjBeforeRegex(str, char, index, regexp);
     }
 
-    private smartAdjBeforeRegex(str: string, char: string, index: number, regexp: RegExp): InteliAdjustment {
+    private smartAdjBeforeRegex(str: string, char: string, index: number, regexp: RegExp): SmartAdjustment {
 
         if (regexp.test(char)) {
             if (index === 0 && str.length !== 1) {
-                if (regexp.test(str[1])) return InteliAdjustment.Before;
+                if (regexp.test(str[1])) return SmartAdjustment.Before;
             } else {
                 if (str[index + 1] && regexp.test(str[index + 1]) && !regexp.test(str[index - 1]))
-                    return InteliAdjustment.Before;
+                    return SmartAdjustment.Before;
             }
         }
-        return InteliAdjustment.Default;
+        return SmartAdjustment.Default;
     }
 
     private indexesOf = (str: string, char: string): CharIndex[] => {
@@ -349,6 +374,12 @@ export class MetaJumper {
         }
 
         let indices = [];
+
+        if (char === '\n') {
+            indices.push(new CharIndex(str.length))
+            return indices;
+        }
+
         let ignoreCase = this.config.jumper.targetIgnoreCase;
         if (this.config.jumper.findAllMode === 'on') {
             for (var i = 0; i < str.length; i++) {
@@ -388,56 +419,62 @@ export class MetaJumper {
         return indices;
     }
 
-    private getExactLocation = async (editor: vscode.TextEditor, models: DecorationModel[]) => {
+    private getExactLocation = async (editorToModelsMap: Map<vscode.TextEditor, DecorationModel[]>) => {
         // show location candidates
-        var decs = this.decorator.create(editor, models);
+        var decs = this.decorator.createAll(editorToModelsMap);
 
         let msg = this.isSelectionMode ? "metaGo: Select To" : "metaGo: Jump To";
         let messageDisposable = vscode.window.setStatusBarMessage(msg, this.config.jumper.timeout);
 
         try {
+            let editor = vscode.window.activeTextEditor ?? vscode.window.visibleTextEditors[0];
             // wait for first code key
-            var value = await new InlineInput(this.config).onKey(this.config.decoration.hide.trigerKey, editor, v => v, 'type the character to goto',
+            var value = await new Input(this.config).onKey(this.config.decoration.hide.trigerKey, editor, v => v, 'type the character to goto',
                 k => { // down
                     if (this.jumpTimeoutId != null) clearTimeout(this.jumpTimeoutId);
-                    this.decorator.hide(editor, decs)
+                    this.decorator.hideAll(decs)
                 }, k => { // up
                     this.jumpTimeoutId = setTimeout(() => { this.jumpTimeoutId = null; this.cancel(); }, this.config.jumper.timeout);
-                    this.decorator.show(editor, decs);
+                    this.decorator.showAll(decs);
                 }, k => {
                 }
             );
 
-            this.decorator.remove(editor);
+            this.decorator.removeAll(decs);
             if (!value) throw new Error('no key code input')
 
-            if (value === '\n') {
-                let model = models.find(model => model.indexInModels === 0);
-                if (model) {
-                    this.currentFindIndex = 0;
-                    return [model]
-                }
-            }
-            else if (value === ' ') {
-                let model = models.find(model => model.indexInModels === 1);
-                if (model) {
-                    this.currentFindIndex = 1;
-                    return [model]
-                }
-            }
+            // if (value === '\n') {
+            //     let model = models.find(model => model.indexInModels === 0);
+            //     if (model) {
+            //         this.currentFindIndex = 0;
+            //         return [model]
+            //     }
+            // }
+            // else if (value === ' ') {
+            //     let model = models.find(model => model.indexInModels === 1);
+            //     if (model) {
+            //         this.currentFindIndex = 1;
+            //         return [model]
+            //     }
+            // }
 
-            // filter location candidates
-            models = models.filter(model => {
-                if(model.code[0] && model.code[0].toLowerCase() === value.toLowerCase()){
-                    model.code = model.code.substring(1)
-                    return true;
-                }
-                return false;
+            editorToModelsMap.forEach((models, editor) => {
+                // filter location candidates
+                models = models.filter(model => {
+                    if (model.code[0] && model.code[0].toLowerCase() === value.toLowerCase()) {
+                        model.code = model.code.substring(1)
+                        return true;
+                    }
+                    return false;
+                })
+
+                if (models.length == 0) editorToModelsMap.delete(editor)
+                else editorToModelsMap.set(editor, models)
             })
-            return models
+            return editorToModelsMap
 
         } catch (e) {
-            this.decorator.remove(editor);
+            this.decorator.removeAll(decs);
             throw e;
         } finally {
             messageDisposable.dispose();